Move kubeconfig keychain pkg under the cmd pkg
Recently, filesystem pkg is refactored and is now agnostic about the auth
implementation (it is abstructed by containerd's RegistryHosts). Because of this
change, `keychain` pkg isn't directly used by that filesystem pkg and is only
used by `cmd/containerd-stargz-grpc` pkg for implementing kubeconfig-based
RegistryHost.

This commit moves the `keychain` pkg under the pkg that consumes this.
