+++ cmd/containerd-stargz-grpc/keychain/kubeconfig.go
