+++ cmd/containerd-stargz-grpc/main.go
@@ -30,9 +30,9 @@ import (
 	"github.com/containerd/containerd/contrib/snapshotservice"
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/containerd/remotes/docker"
+	"github.com/containerd/stargz-snapshotter/cmd/containerd-stargz-grpc/keychain"
 	snbase "github.com/containerd/stargz-snapshotter/snapshot"
 	"github.com/containerd/stargz-snapshotter/stargz"
-	"github.com/containerd/stargz-snapshotter/stargz/keychain"
 	"github.com/containerd/stargz-snapshotter/stargz/source"
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/google/go-containerregistry/pkg/name"
